Moved utility test to tests directory.
