+++ python/moose/utils.py
@@ -1,4 +1,4 @@
-# -* coding: utf-8 -*-
+# -*- coding: utf-8 -*-
 # Utility functions for moose.
 
 from __future__ import print_function, division, absolute_import
@@ -336,7 +336,7 @@ def autoposition(root):
 def loadModel(filename, target,method='ee'):
     moose.loadModel(filename,target)
     moose.mooseaddChemSolver(target,method)
-	
+
 def readcell_scrambled(filename, target, method='ee'):
     """A special version for handling cases where a .p file has a line
     with specified parent yet to be defined.
@@ -842,180 +842,3 @@ def get_child_Mstring(mooseobject,mstring):
             return child
     return None
 
-# Note: This function is also moved to helper.moose_methods
-def connect_CaConc(compartment_list, temperature=None):
-    """ Connect the Ca pools and channels within each of the compartments in compartment_list
-     Ca channels should have a child Mstring named 'ion' with value set in MOOSE.
-     Ca dependent channels like KCa should have a child Mstring called 'ionDependency' with value set in MOOSE.
-     Call this only after instantiating cell so that all channels and pools have been created. """
-    for compartment in compartment_list:
-        caconc = None
-        for child in compartment.children:
-            neutralwrap = moose.Neutral(child)
-            if neutralwrap.className == 'CaConc':
-                caconc = moose.CaConc(child)
-                break
-        if caconc is not None:
-            child = get_child_Mstring(caconc,'phi')
-            if child is not None:
-                caconc.B = float(child.value) # B = phi by definition -- see neuroml 1.8.1 defn
-            else:
-                ## B has to be set for caconc based on thickness of Ca shell and compartment l and dia,
-                ## OR based on the Mstring phi under CaConc path.
-                ## I am using a translation from Neuron for mitral cell, hence this method.
-                ## In Genesis, gmax / (surfacearea*thick) is set as value of B!
-                caconc.B = 1 / (2*FARADAY) / \
-                    (math.pi*compartment.diameter*compartment.length * caconc.thick)
-            for child in compartment.children:
-                neutralwrap = moose.Neutral(child)
-                if neutralwrap.className == 'HHChannel':
-                    channel = moose.HHChannel(child)
-                    ## If child Mstring 'ion' is present and is Ca, connect channel current to caconc
-                    for childid in channel.children:
-                        # in async13, gates which have not been created still 'exist'
-                        # i.e. show up as a child, but cannot be wrapped.
-                        try:
-                            child = moose.element(childid)
-                            if child.className=='Mstring':
-                                child = moose.Mstring(child)
-                                if child.name=='ion':
-                                    if child.value in ['Ca','ca']:
-                                        moose.connect(channel,'IkOut',caconc,'current')
-                                        #print 'Connected IkOut of',channel.path,'to current of',caconc.path
-                                ## temperature is used only by Nernst part here...
-                                if child.name=='nernst_str':
-                                    nernst = moose.Nernst(channel.path+'/nernst')
-                                    nernst_params = child.value.split(',')
-                                    nernst.Cout = float(nernst_params[0])
-                                    nernst.valence = float(nernst_params[1])
-                                    nernst.Temperature = temperature
-                                    moose.connect(nernst,'Eout',channel,'setEk')
-                                    moose.connect(caconc,'concOut',nernst,'ci')
-                                    #print 'Connected Nernst',nernst.path
-                        except TypeError:
-                            pass
-
-                if neutralwrap.className == 'HHChannel2D':
-                    channel = moose.HHChannel2D(child)
-                    ## If child Mstring 'ionDependency' is present, connect caconc Ca conc to channel
-                    for childid in channel.children:
-                        # in async13, gates which have not been created still 'exist'
-                        # i.e. show up as a child, but cannot be wrapped.
-                        try:
-                            child = moose.element(childid)
-                            if child.className=='Mstring' and child.name=='ionDependency':
-                                child = moose.Mstring(child)
-                                if child.value in ['Ca','ca']:
-                                    moose.connect(caconc,'concOut',channel,'concen')
-                                    #print 'Connected concOut of',caconc.path,'to concen of',channel.path
-                        except TypeError:
-                            pass
-
-############# added by Aditya Gilra -- end ################
-import uuid
-import unittest
-import sys
-from io import StringIO as _sio
-
-class _TestMooseUtils(unittest.TestCase):
-    def test_printtree(self):
-        s = moose.Neutral('/cell')
-        soma = moose.Neutral('%s/soma'% (s.path))
-        d1 = moose.Neutral('%s/d1'% (soma.path))
-        d2 = moose.Neutral('%s/d2'% (soma.path))
-        d3 = moose.Neutral('%s/d3'% (d1.path))
-        d4 = moose.Neutral('%s/d4'% (d1.path))
-        d5 = moose.Neutral('%s/d5'% (s.path))
-        orig_stdout = sys.stdout
-        sys.stdout = _sio()
-        printtree(s)
-        expected = """
-cell
-|
-|__ soma
-|  |
-|  |__ d1
-|  |  |
-|  |  |__ d3
-|  |  |
-|  |  |__ d4
-|  |
-|  |__ d2
-|
-|__ d5
-"""
-        self.assertEqual(sys.stdout.getvalue(), expected)
-        sys.stdout = _sio()
-        s1 = moose.Neutral('cell1')
-        c1 = moose.Neutral('%s/c1' % (s1.path))
-        c2 = moose.Neutral('%s/c2' % (c1.path))
-        c3 = moose.Neutral('%s/c3' % (c1.path))
-        c4 = moose.Neutral('%s/c4' % (c2.path))
-        c5 = moose.Neutral('%s/c5' % (c3.path))
-        c6 = moose.Neutral('%s/c6' % (c3.path))
-        c7 = moose.Neutral('%s/c7' % (c4.path))
-        c8 = moose.Neutral('%s/c8' % (c5.path))
-        printtree(s1)
-        expected1 = """
-cell1
-|
-|__ c1
-   |
-   |__ c2
-   |  |
-   |  |__ c4
-   |     |
-   |     |__ c7
-   |
-   |__ c3
-      |
-      |__ c5
-      |  |
-      |  |__ c8
-      |
-      |__ c6
-"""
-        self.assertEqual(sys.stdout.getvalue(), expected1)
-
-    def test_autoposition(self):
-        """Simple check for automatic generation of positions.
-
-        A spherical soma is created with 20 um diameter. A 100
-        compartment cable is created attached to it with each
-        compartment of length 100 um.
-
-        """
-        testid = 'test%s' % (uuid.uuid4())
-        container = moose.Neutral('/test')
-        model = moose.Neuron('/test/%s' % (testid))
-        soma = moose.Compartment('%s/soma' % (model.path))
-        soma.diameter = 20e-6
-        soma.length = 0.0
-        parent = soma
-        comps = []
-        for ii in range(100):
-            comp = moose.Compartment('%s/comp_%d' % (model.path, ii))
-            comp.diameter = 10e-6
-            comp.length = 100e-6
-            moose.connect(parent, 'raxial', comp, 'axial')
-            comps.append(comp)
-            parent = comp
-        soma = autoposition(model)
-        sigfig = 8
-        self.assertAlmostEqual(soma.x0, 0.0, sigfig)
-        self.assertAlmostEqual(soma.y0, 0.0, sigfig)
-        self.assertAlmostEqual(soma.z0, 0.0, sigfig)
-        self.assertAlmostEqual(soma.x, 0.0, sigfig)
-        self.assertAlmostEqual(soma.y, 0.0, sigfig)
-        self.assertAlmostEqual(soma.z, soma.diameter/2.0, sigfig)
-        for ii, comp in enumerate(comps):
-            print(comp.path, ii)
-            self.assertAlmostEqual(comp.x0, 0, sigfig)
-            self.assertAlmostEqual(comp.y0, 0.0, sigfig)
-            self.assertAlmostEqual(comp.z0, soma.diameter/2.0 + ii * 100e-6, sigfig)
-            self.assertAlmostEqual(comp.x, 0.0, sigfig)
-            self.assertAlmostEqual(comp.y, 0.0, sigfig)
-            self.assertAlmostEqual(comp.z, soma.diameter/2.0 + (ii + 1) * 100e-6, sigfig)
-
-if __name__ == "__main__": # test printtree
-    unittest.main()
